fixed all pages to work for links correctly, redid app.js and put all previous app.js code into Home page, made new pages for each part of app
